[#20] Refactor, extract random to utils
